fix(loading): moved token logic to homepage
